Support __getitem__ for dataframes
